Make SVG QR codes scaleable
by using the viewBox parameter instead of directly setting width and height, the svg can scale cleanly to any size display. 
This backwards compatible change would default to the same behavior, but set either a defined pixel size of the users choice, or 'auto'/-1 to not include the width/height, making the svg scale to 100% of the parent width.